requirements.txt
absl-py==0.7.0
aircv==1.4.6
alabaster==0.7.10
altgraph==0.16.1
anaconda-client==1.6.14
anaconda-navigator==1.9.6
anaconda-project==0.8.2
apkutils==0.5.2
asn1crypto==0.24.0
astor==0.7.1
astroid==1.6.3
astropy==3.0.2
attrs==18.1.0
atx==1.2.1.dev8
Automat==0.7.0
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
Bottleneck==1.2.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
chatterbot-corpus==1.1.4
cigam==0.0.3
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
concurrent-log-handler==0.9.12
conda==4.3.30
conda-build==3.10.5
conda-verify==2.0.0
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.2.2
cssselect==1.0.3
cx-Oracle==6.4.1
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dask==0.17.5
datashape==0.5.4
DBUtils==1.3
decorator==4.3.0
distributed==1.21.8
Django==1.8.18
docopt==0.6.2
docutils==0.14
dwebsocket==0.4.2
emoji==0.5.0
entrypoints==0.2.3
et-xmlfile==1.0.1
facebook-wda==0.2.2.dev7
fastcache==1.0.2
filelock==3.0.4
fire==0.1.3
Flask==1.0.2
Flask-Cors==3.0.4
future==0.16.0
futures==3.1.1
gast==0.2.2
gevent==1.3.0
glob2==0.6
greenlet==0.4.13
grpcio==1.18.0
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
humanize==0.5.1
hyperlink==18.0.0
idna==2.6
imageio==2.3.0
imagesize==1.0.0
imutils==0.4.6
incremental==17.5.0
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itchat==1.3.10
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
llvmlite==0.23.1
locket==0.2.0
logzero==1.5.0
lxml==4.2.1
macholib==1.11
maproxy==0.0.12
Markdown==3.0.1
MarkupSafe==1.0
mathparse==0.1.2
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.5.6
multipledispatch==0.5.0
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.14.3
numpydoc==0.8.0
oauthlib==2.1.0
odo==0.5.1
olefile==0.45.1
opencv-contrib-python==3.4.4.19
opencv-python==3.4.1.15
openpyxl==2.5.3
packaging==17.1
pandas==0.23.0
pandocfilters==1.4.2
paramiko==2.4.2
parsel==1.5.1
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pefile==2018.8.8
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
pluggy==0.6.0
ply==3.11
progress==1.4
prompt-toolkit==1.0.15
protobuf==3.6.1
psutil==5.4.5
pure-python-adb==0.1.6.dev0
py==1.5.3
pyasn1==0.4.4
pyasn1-modules==0.2.4
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.1
PyDispatcher==2.0.5
pyelftools==0.25
pyflakes==1.6.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyInstaller==3.4
pylint==1.8.4
pymongo==3.7.2
pymssql==2.1.3
PyMySQL==0.9.2
PyNaCl==1.3.0
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pypng==0.0.18
Pypubsub==4.0.0
PyQRCode==1.2.1
PySocks==1.6.8
pysqlcipher3==1.0.2
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.6.1
python-docx==0.8.7
python-twitter==3.4.2
pytz==2018.4
PyWavelets==0.5.2
pywin32==224
pywin32-ctypes==0.2.0
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
queuelib==1.5.0
requests==2.18.4
requests-oauthlib==1.0.0
retry==0.9.2
rope==0.10.7
ruamel-yaml==0.15.35
scapy==2.4.0
schedule==0.5.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
Scrapy==1.6.0
seaborn==0.8.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tblib==1.3.2
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.0.2
tqdm==4.28.1
traitlets==4.3.2
Twisted==18.9.0
typing==3.6.4
uiautomator2==0.1.11.dev1
unicodecsv==0.14.1
urllib3==1.22
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
weditor==0.1.1
Werkzeug==0.14.1
wget==3.2
whichcraft==0.4.1
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
wxPython==4.0.3
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
xmltodict==0.11.0
zict==0.1.3
zope.interface==4.6.0
